Add files via upload
Disables Windows Defender in a way that is relatively easy to reverse. Navigates through settings and add the C:\ drive as an exclusion. Can be modified to add any specific folder or file as an exclusion by editing the last STING line. Tested on Windows 10, not sure about Windows 11. only works if nothing is disabled in defender settings.